pyproject.toml
[tool.poetry]
name = "speech-synthesis-spotify"
version = "0.1.0"
description = ""
authors = ["malyvsen <5940672+malyvsen@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.7"
guildai = "^0.7.2"
SoundFile = "^0.10.3"
Keras = "^2.4.3"
opencv-python = "^4.5.1"
tensorflow = "^2.4.1"
scikit-image = "^0.18.1"
praatio = "^4.2.1"
Pillow = "^8.1.0"
librosa = "^0.8.0"
tqdm = "^4.57.0"
numpy = "^1.19.2"
scipy = "^1.6.1"
g2p-en = "^2.1.0"
matplotlib = "3.2"
praat-parselmouth = "^0.4.0"

[tool.poetry.dev-dependencies]
scipy = "^1.6.1"
black = {version = "^20.8b1", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
